requirements.txt
Django==1.5.1
requests==1.2.0
PyYAML
South
gunicorn
dj-database-url